test: Clean up tests
- Improve testing utils and place them in appropriate separate file
- Break tests up into separate files for slightly easier navigation
- Extend expect with custom matchers for checking video state
