--- tests/jest.init.js
@@ -1,3 +1,0 @@
-// Configures test environment with polyfills before the test framework is installed in the environment
-import 'core-js/stable';
-import 'regenerator-runtime/runtime';
